requirements.txt
accelerate==0.21.0
deepspeed==0.10.1
flash-attn==2.1.1
jsonlines==3.1.0
nltk==3.8.1
numpy==1.24.4
openai==0.27.8
protobuf==4.24.0
safetensors==0.3.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
spacy==2.2.4
tiktoken==0.5.1
tokenizers==0.15.0
# torch==2.0.1+cu118  # 不再需要，因為我們用 pytorch/pytorch 官方映像
tqdm==4.66.1
transformers==4.35.2
uvicorn==0.23.2
vllm==0.2.1.post1

# 新增以下套件
fastapi==0.103.1
python-dotenv==1.0.0